+++ src/main/scala/cromwell/engine/db/DataAccess.scala
@@ -8,7 +8,7 @@ import cromwell.engine.{WorkflowOutputs, _}
 import cromwell.webservice.{CallCachingParameters, WorkflowQueryParameters, WorkflowQueryResponse}
 import wdl4s.{CallInputs, _}
 
-import scala.concurrent.Future
+import scala.concurrent.{ExecutionContext, Future}
 
 object DataAccess {
   val globalDataAccess: DataAccess = new slick.SlickDataAccess()
@@ -22,89 +22,109 @@ trait DataAccess extends AutoCloseable {
   def createWorkflow(workflowDescriptor: WorkflowDescriptor,
                      workflowInputs: Traversable[SymbolStoreEntry],
                      calls: Traversable[Scope],
-                     backend: Backend): Future[Unit]
+                     backend: Backend)(implicit ec: ExecutionContext): Future[Unit]
 
-  def getWorkflowState(workflowId: WorkflowId): Future[Option[WorkflowState]]
+  def getWorkflowState(workflowId: WorkflowId)(implicit ec: ExecutionContext): Future[Option[WorkflowState]]
 
-  def getWorkflow(workflowId: WorkflowId): Future[WorkflowDescriptor]
+  def getWorkflow(workflowId: WorkflowId)(implicit ec: ExecutionContext): Future[WorkflowDescriptor]
 
-  def getWorkflow(workflowExecutionId: Int): Future[WorkflowDescriptor]
+  def getWorkflow(workflowExecutionId: Int)(implicit ec: ExecutionContext): Future[WorkflowDescriptor]
 
-  def getWorkflowsByState(states: Traversable[WorkflowState]): Future[Traversable[WorkflowDescriptor]]
+  def getWorkflowsByState(states: Traversable[WorkflowState])
+                         (implicit ec: ExecutionContext): Future[Traversable[WorkflowDescriptor]]
 
-  def getExecutionBackendInfo(workflowId: WorkflowId, call: Call, attempt: Int): Future[CallBackendInfo]
+  def getExecutionBackendInfo(workflowId: WorkflowId, call: Call, attempt: Int)
+                             (implicit ec: ExecutionContext): Future[CallBackendInfo]
 
-  def updateExecutionBackendInfo(workflowId: WorkflowId, callKey: BackendCallKey, backendInfo: CallBackendInfo): Future[Unit]
+  def updateExecutionBackendInfo(workflowId: WorkflowId, callKey: BackendCallKey, backendInfo: CallBackendInfo)
+                                (implicit ec: ExecutionContext): Future[Unit]
 
-  def updateWorkflowState(workflowId: WorkflowId, workflowState: WorkflowState): Future[Unit]
+  def updateWorkflowState(workflowId: WorkflowId, workflowState: WorkflowState)
+                         (implicit ec: ExecutionContext): Future[Unit]
 
-  def getAllSymbolStoreEntries(workflowId: WorkflowId): Future[Traversable[SymbolStoreEntry]]
+  def getAllSymbolStoreEntries(workflowId: WorkflowId)
+                              (implicit ec: ExecutionContext): Future[Traversable[SymbolStoreEntry]]
 
   // TODO needed to support compatibility with current code, this seems like an inefficient way of getting
   // TODO workflow outputs.
   /** Returns all outputs for this workflowId */
-  def getWorkflowOutputs(workflowId: WorkflowId): Future[Traversable[SymbolStoreEntry]]
+  def getWorkflowOutputs(workflowId: WorkflowId)(implicit ec: ExecutionContext): Future[Traversable[SymbolStoreEntry]]
 
-  def getAllOutputs(workflowId: WorkflowId): Future[Traversable[SymbolStoreEntry]]
+  def getAllOutputs(workflowId: WorkflowId)(implicit ec: ExecutionContext): Future[Traversable[SymbolStoreEntry]]
 
-  def getAllInputs(workflowId: WorkflowId): Future[Traversable[SymbolStoreEntry]]
+  def getAllInputs(workflowId: WorkflowId)(implicit ec: ExecutionContext): Future[Traversable[SymbolStoreEntry]]
 
   /** Get all outputs for the scope of this call. */
-  def getOutputs(workflowId: WorkflowId, key: ExecutionDatabaseKey): Future[Traversable[SymbolStoreEntry]]
+  def getOutputs(workflowId: WorkflowId, key: ExecutionDatabaseKey)
+                (implicit ec: ExecutionContext): Future[Traversable[SymbolStoreEntry]]
 
   /** Get all inputs for the scope of this call. */
-  def getInputs(id: WorkflowId, call: Call): Future[Traversable[SymbolStoreEntry]]
+  def getInputs(id: WorkflowId, call: Call)(implicit ec: ExecutionContext): Future[Traversable[SymbolStoreEntry]]
 
   /** Should fail if a value is already set.  The keys in the Map are locally qualified names. */
-  def setOutputs(workflowId: WorkflowId, key: OutputKey, callOutputs: WorkflowOutputs, workflowOutputFqns: Seq[ReportableSymbol]): Future[Unit]
+  def setOutputs(workflowId: WorkflowId, key: OutputKey, callOutputs: WorkflowOutputs,
+                 workflowOutputFqns: Seq[ReportableSymbol])(implicit ec: ExecutionContext): Future[Unit]
 
   /** Updates the existing input symbols to replace expressions with real values **/
-  def updateCallInputs(workflowId: WorkflowId, key: BackendCallKey, callInputs: CallInputs): Future[Int]
+  def updateCallInputs(workflowId: WorkflowId, key: BackendCallKey, callInputs: CallInputs)
+                      (implicit ec: ExecutionContext): Future[Int]
 
-  def setExecutionEvents(workflowId: WorkflowId, callFqn: String, shardIndex: Option[Int], attempt: Int, events: Seq[ExecutionEventEntry]): Future[Unit]
+  def setExecutionEvents(workflowId: WorkflowId, callFqn: String, shardIndex: Option[Int], attempt: Int,
+                         events: Seq[ExecutionEventEntry])(implicit ec: ExecutionContext): Future[Unit]
 
   /** Gets a mapping from call FQN to an execution event entry list */
-  def getAllExecutionEvents(workflowId: WorkflowId): Future[Map[ExecutionDatabaseKey, Seq[ExecutionEventEntry]]]
+  def getAllExecutionEvents(workflowId: WorkflowId)
+                           (implicit ec: ExecutionContext): Future[Map[ExecutionDatabaseKey, Seq[ExecutionEventEntry]]]
 
-  def setStatus(workflowId: WorkflowId, keys: Traversable[ExecutionDatabaseKey], executionStatus: ExecutionStatus): Future[Unit] = {
+  def setStatus(workflowId: WorkflowId, keys: Traversable[ExecutionDatabaseKey], executionStatus: ExecutionStatus)
+               (implicit ec: ExecutionContext): Future[Unit] = {
     setStatus(workflowId, keys, CallStatus(executionStatus, None, None, None))
   }
 
-  def setStatus(workflowId: WorkflowId, keys: Traversable[ExecutionDatabaseKey], callStatus: CallStatus): Future[Unit]
+  def setStatus(workflowId: WorkflowId, keys: Traversable[ExecutionDatabaseKey], callStatus: CallStatus)
+               (implicit ec: ExecutionContext): Future[Unit]
 
-  def getExecutionStatuses(workflowId: WorkflowId): Future[Map[ExecutionDatabaseKey, CallStatus]]
+  def getExecutionStatuses(workflowId: WorkflowId)
+                          (implicit ec: ExecutionContext): Future[Map[ExecutionDatabaseKey, CallStatus]]
 
   /** Return all execution entries for the FQN, including collector and shards if any */
-  def getExecutionStatuses(workflowId: WorkflowId, fqn: FullyQualifiedName): Future[Map[ExecutionDatabaseKey, CallStatus]]
+  def getExecutionStatuses(workflowId: WorkflowId, fqn: FullyQualifiedName)
+                          (implicit ec: ExecutionContext): Future[Map[ExecutionDatabaseKey, CallStatus]]
 
-  def getExecutionStatus(workflowId: WorkflowId, key: ExecutionDatabaseKey): Future[Option[CallStatus]]
+  def getExecutionStatus(workflowId: WorkflowId, key: ExecutionDatabaseKey)
+                        (implicit ec: ExecutionContext): Future[Option[CallStatus]]
 
-  def insertCalls(workflowId: WorkflowId, keys: Traversable[ExecutionStoreKey], backend: Backend): Future[Unit]
+  def insertCalls(workflowId: WorkflowId, keys: Traversable[ExecutionStoreKey], backend: Backend)
+                 (implicit ec: ExecutionContext): Future[Unit]
 
-  def getExecutions(id: WorkflowId): Future[Traversable[Execution]]
+  def getExecutions(id: WorkflowId)(implicit ec: ExecutionContext): Future[Traversable[Execution]]
 
-  def getExecutionsForRestart(id: WorkflowId): Future[Traversable[Execution]]
+  def getExecutionsForRestart(id: WorkflowId)(implicit ec: ExecutionContext): Future[Traversable[Execution]]
 
-  def getExecutionsWithResuableResultsByHash(hash: String): Future[Traversable[Execution]]
+  def getExecutionsWithResuableResultsByHash(hash: String)
+                                            (implicit ec: ExecutionContext): Future[Traversable[Execution]]
 
   /** Fetch the workflow having the specified `WorkflowId`. */
-  def getWorkflowExecution(workflowId: WorkflowId): Future[WorkflowExecution]
+  def getWorkflowExecution(workflowId: WorkflowId)(implicit ec: ExecutionContext): Future[WorkflowExecution]
 
-  def getWorkflowExecutionAux(id: WorkflowId): Future[WorkflowExecutionAux]
+  def getWorkflowExecutionAux(id: WorkflowId)(implicit ec: ExecutionContext): Future[WorkflowExecutionAux]
 
-  def jesJobInfo(id: WorkflowId): Future[Map[ExecutionDatabaseKey, JesJob]]
+  def jesJobInfo(id: WorkflowId)(implicit ec: ExecutionContext): Future[Map[ExecutionDatabaseKey, JesJob]]
 
-  def localJobInfo(id: WorkflowId): Future[Map[ExecutionDatabaseKey, LocalJob]]
+  def localJobInfo(id: WorkflowId)(implicit ec: ExecutionContext): Future[Map[ExecutionDatabaseKey, LocalJob]]
 
-  def sgeJobInfo(id: WorkflowId): Future[Map[ExecutionDatabaseKey, SgeJob]]
+  def sgeJobInfo(id: WorkflowId)(implicit ec: ExecutionContext): Future[Map[ExecutionDatabaseKey, SgeJob]]
 
-  def updateWorkflowOptions(workflowId: WorkflowId, workflowOptionsJson: String): Future[Unit]
+  def updateWorkflowOptions(workflowId: WorkflowId, workflowOptionsJson: String)
+                           (implicit ec: ExecutionContext): Future[Unit]
 
-  def resetNonResumableJesExecutions(workflowId: WorkflowId): Future[Unit]
+  def resetNonResumableJesExecutions(workflowId: WorkflowId)(implicit ec: ExecutionContext): Future[Unit]
 
-  def findResumableJesExecutions(workflowId: WorkflowId): Future[Map[ExecutionDatabaseKey, JobKey]]
+  def findResumableJesExecutions(workflowId: WorkflowId)
+                                (implicit ec: ExecutionContext): Future[Map[ExecutionDatabaseKey, JobKey]]
 
-  def queryWorkflows(queryParameters: WorkflowQueryParameters): Future[WorkflowQueryResponse]
+  def queryWorkflows(queryParameters: WorkflowQueryParameters)
+                    (implicit ec: ExecutionContext): Future[WorkflowQueryResponse]
 
-  def updateCallCaching(cachingParameters: CallCachingParameters): Future[Int]
+  def updateCallCaching(cachingParameters: CallCachingParameters)(implicit ec: ExecutionContext): Future[Int]
 }